feat: adding relation accept and change from 32 byte int to 64 byte int

+++ internal/adapters/inbound/rest/relation/dto/create_relation_dto.go
@@ -1,7 +1,7 @@
 package relation_dto
 
 type CreateRelationDTO struct {
-	FromId       int32
-	TargetId     int32
+	FromId       int64
+	TargetId     int64
 	RelationType string
 }
