+++ internal/domain/entities/profile.go
@@ -3,11 +3,11 @@ package entities
 import "errors"
 
 type Profile struct {
-	ConnectId int32
+	ConnectId int64
 	Name      string
 }
 
-func NewProfile(id int32, name string) (*Profile, error) {
+func NewProfile(id int64, name string) (*Profile, error) {
 	if name == "" {
 		return nil, errors.New("name cannot be empty")
 	}
