+++ internal/domain/entities/profile_test.go
@@ -6,7 +6,7 @@ import (
 
 func TestShouldDeclareProfileWithAllValidProperties(t *testing.T) {
 	name := "Akira"
-	var id int32 = 1
+	var id int64 = 1
 	profile, err := NewProfile(id, name)
 	if profile == nil || err != nil {
 		t.Error("Should name and id be valid and create the profile")
@@ -15,7 +15,7 @@ func TestShouldDeclareProfileWithAllValidProperties(t *testing.T) {
 
 func TestShouldReturnErrorWithInvalidName(t *testing.T) {
 	name := ""
-	var id int32 = 1
+	var id int64 = 1
 	profile, err := NewProfile(id, name)
 	if profile != nil || err.Error() != "name cannot be empty" {
 		t.Error("Should return name is invalid")
@@ -24,7 +24,7 @@ func TestShouldReturnErrorWithInvalidName(t *testing.T) {
 
 func TestShouldReturnErrorWithInvalidId(t *testing.T) {
 	name := "akira"
-	var id int32 = 0
+	var id int64 = 0
 	profile, err := NewProfile(id, name)
 	if profile != nil || err.Error() != "id cannot be 0" {
 		t.Error("Should return id is invalid")
