+++ internal/domain/entities/relation_test.go
@@ -3,8 +3,8 @@ package entities
 import "testing"
 
 func TestShouldRelationBeCreatedWhenAllParametersAreValid(t *testing.T) {
-	var id int32 = 1
-	var toId int32 = 2
+	var id int64 = 1
+	var toId int64 = 2
 	var relationType RelationType = RelationFriend
 	relation, err := NewRelation(id, toId, relationType, RelationStatusPending)
 	if err != nil || relation == nil {
@@ -13,8 +13,8 @@ func TestShouldRelationBeCreatedWhenAllParametersAreValid(t *testing.T) {
 }
 
 func TestShouldThrowErrorWhenToIdIsEqualFromID(t *testing.T) {
-	var id int32 = 1
-	var toId int32 = 1
+	var id int64 = 1
+	var toId int64 = 1
 	var relationType RelationType = RelationFriend
 	relation, err := NewRelation(id, toId, relationType, RelationStatusRejected)
 	if err.Error() != "cannot create relation with self" || relation != nil {
@@ -23,8 +23,8 @@ func TestShouldThrowErrorWhenToIdIsEqualFromID(t *testing.T) {
 }
 
 func TestShouldThrowErrorWhenRelationTypeIsNone(t *testing.T) {
-	var id int32 = 2
-	var toId int32 = 1
+	var id int64 = 2
+	var toId int64 = 1
 	var relationType RelationType
 	relation, err := NewRelation(id, toId, relationType, RelationStatusAccepted)
 	if err.Error() != "relation type is required" || relation != nil {
