+++ internal/domain/service/profile_service.go
@@ -19,8 +19,8 @@ func (s *ProfileService) CreateProfile(ctx context.Context, profile *entities.Pr
 	return s.repository.CreateProfile(ctx, profile)
 }
 
-func (s *ProfileService) DeleteProfile(ctx context.Context, id int32) error {
-	if id == int32(0) {
+func (s *ProfileService) DeleteProfile(ctx context.Context, id int64) error {
+	if id == int64(0) {
 		return errors.New("id to delete cannot be 0")
 	}
 	return s.repository.DeleteProfile(ctx, id)
