+++ internal/domain/service/profile_service_test.go
@@ -20,7 +20,7 @@ func (m *MockProfileRepository) CreateProfile(ctx context.Context, profile *enti
 	return args.Get(0).(entities.Profile), args.Error(1)
 }
 
-func (m *MockProfileRepository) DeleteProfile(ctx context.Context, id int32) error {
+func (m *MockProfileRepository) DeleteProfile(ctx context.Context, id int64) error {
 	args := m.Called(ctx, id)
 
 	return args.Error(0)
@@ -82,23 +82,23 @@ func TestProfileServic_DeleteProfile(t *testing.T) {
 	t.Run("should delete with success when profile is ok", func(t *testing.T) {
 		mockRepo := new(MockProfileRepository)
 		service := CreateNewProfileService(mockRepo)
-		mockRepo.On("DeleteProfile", ctx, int32(1)).Return(nil)
-		err := service.DeleteProfile(ctx, int32(1))
+		mockRepo.On("DeleteProfile", ctx, int64(1)).Return(nil)
+		err := service.DeleteProfile(ctx, int64(1))
 		assert.Empty(t, err)
 	})
 
 	t.Run("should return error when profile return error", func(t *testing.T) {
 		mockRepo := new(MockProfileRepository)
 		service := CreateNewProfileService(mockRepo)
-		mockRepo.On("DeleteProfile", ctx, int32(1)).Return(errors.New("Some error"))
-		err := service.DeleteProfile(ctx, int32(1))
+		mockRepo.On("DeleteProfile", ctx, int64(1)).Return(errors.New("Some error"))
+		err := service.DeleteProfile(ctx, int64(1))
 		assert.Equal(t, err.Error(), "Some error")
 	})
 
 	t.Run("should return error when id is 0", func(t *testing.T) {
 		mockRepo := new(MockProfileRepository)
 		service := CreateNewProfileService(mockRepo)
-		err := service.DeleteProfile(ctx, int32(0))
+		err := service.DeleteProfile(ctx, int64(0))
 		assert.Equal(t, err.Error(), "id to delete cannot be 0")
 	})
 }
